backend/auth_service_v2/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3
ratelimit==2.2.1

# Logging
loguru==0.7.2 